Refactor friends service and auth0 cleanup
